Bring some of the common steps together in a spec support module
