+++ spec/support/system/common_steps.rb
@@ -1,0 +1,51 @@
+module CommonSteps
+  def given_the_service_is_open
+    FeatureFlags::FeatureFlag.activate(:service_open)
+  end
+
+  def and_i_am_signed_in
+    @user = create(:user)
+    sign_in(@user)
+  end
+
+  def and_the_employer_form_feature_is_active
+    FeatureFlags::FeatureFlag.activate(:employer_form)
+  end
+
+  def and_the_user_accounts_feature_is_active
+    FeatureFlags::FeatureFlag.activate(:user_accounts)
+  end
+
+  def and_the_eligbility_screener_feature_is_active
+    FeatureFlags::FeatureFlag.activate(:eligibility_screener)
+  end
+
+  def and_i_have_an_existing_referral
+    @referral = create(:referral, user: @user)
+  end
+  alias_method :when_i_have_an_existing_referral,
+               :and_i_have_an_existing_referral
+
+  def and_i_visit_the_referral
+    visit edit_referral_path(@referral)
+  end
+  alias_method :when_i_visit_the_referral, :and_i_visit_the_referral
+
+  def then_i_see_the_referral_summary
+    expect(page).to have_current_path(edit_referral_path(@referral))
+    expect(page).to have_title(
+      "Refer serious misconduct by a teacher in England"
+    )
+    expect(page).to have_content("Your allegation of serious misconduct")
+  end
+
+  def when_i_click_save_and_continue
+    click_on "Save and continue"
+  end
+  alias_method :and_i_click_save_and_continue, :when_i_click_save_and_continue
+
+  def when_i_click_back
+    click_on "Back"
+  end
+  alias_method :and_i_click_back, :when_i_click_back
+end
