+++ spec/system/referrals/user_adds_an_allegation_spec.rb
@@ -2,12 +2,15 @@
 require "rails_helper"
 
 RSpec.feature "Allegation", type: :system do
+  include CommonSteps
+
   scenario "User adds an allegation to a referral" do
     given_the_service_is_open
     and_i_am_signed_in
     and_the_employer_form_feature_is_active
     and_the_user_accounts_feature_is_active
-    and_i_visit_a_referral
+    and_i_have_an_existing_referral
+    and_i_visit_the_referral
     then_i_see_the_referral_summary
 
     when_i_edit_the_allegation
@@ -38,32 +41,6 @@ RSpec.feature "Allegation", type: :system do
 
   private
 
-  def given_the_service_is_open
-    FeatureFlags::FeatureFlag.activate(:service_open)
-  end
-
-  def and_i_am_signed_in
-    @user = create(:user)
-    sign_in(@user)
-  end
-
-  def and_the_employer_form_feature_is_active
-    FeatureFlags::FeatureFlag.activate(:employer_form)
-  end
-
-  def and_the_user_accounts_feature_is_active
-    FeatureFlags::FeatureFlag.activate(:user_accounts)
-  end
-
-  def and_i_visit_a_referral
-    @referral = create(:referral, user: @user)
-    visit edit_referral_path(@referral)
-  end
-
-  def then_i_see_the_referral_summary
-    expect(page).to have_content("Your allegation of serious misconduct")
-  end
-
   def when_i_edit_the_allegation
     within(all(".app-task-list__section")[2]) do
       click_on "Details of the allegation"
@@ -75,11 +52,6 @@ RSpec.feature "Allegation", type: :system do
       "How do you want to tell us about your allegation?"
     )
   end
-
-  def when_i_click_save_and_continue
-    click_on "Save and continue"
-  end
-  alias_method :and_i_click_save_and_continue, :when_i_click_save_and_continue
 
   def then_i_see_allegation_form_validation_errors
     expect(page).to have_content(
