+++ spec/system/referrals/user_adds_evidence_spec.rb
@@ -2,12 +2,15 @@
 require "rails_helper"
 
 RSpec.feature "Evidence", type: :system do
+  include CommonSteps
+
   scenario "User adds evidence to a referral" do
     given_the_service_is_open
     and_i_am_signed_in
     and_the_employer_form_feature_is_active
     and_the_user_accounts_feature_is_active
-    and_i_visit_a_referral
+    and_i_have_an_existing_referral
+    and_i_visit_the_referral
     then_i_see_the_referral_summary
 
     when_i_edit_the_evidence
@@ -71,33 +74,6 @@ RSpec.feature "Evidence", type: :system do
 
   private
 
-  def given_the_service_is_open
-    FeatureFlags::FeatureFlag.activate(:service_open)
-  end
-
-  def and_i_am_signed_in
-    @user = create(:user)
-    sign_in(@user)
-  end
-
-  def and_the_employer_form_feature_is_active
-    FeatureFlags::FeatureFlag.activate(:employer_form)
-  end
-
-  def and_the_user_accounts_feature_is_active
-    FeatureFlags::FeatureFlag.activate(:user_accounts)
-  end
-
-  def and_i_visit_a_referral
-    @referral = create(:referral, user: @user)
-
-    visit edit_referral_path(@referral)
-  end
-
-  def then_i_see_the_referral_summary
-    expect(page).to have_content("Your allegation of serious misconduct")
-  end
-
   def when_i_edit_the_evidence
     within(all(".app-task-list__section")[2]) do
       click_on "Evidence and supporting information"
@@ -107,11 +83,6 @@ RSpec.feature "Evidence", type: :system do
   def then_i_am_asked_if_i_have_evidence_to_upload
     expect(page).to have_content("Evidence and supporting information")
   end
-
-  def when_i_click_save_and_continue
-    click_on "Save and continue"
-  end
-  alias_method :and_i_click_save_and_continue, :when_i_click_save_and_continue
 
   def then_i_see_evidence_start_form_validation_errors
     expect(page).to have_content("Tell us if you have evidence to upload")
@@ -175,10 +146,6 @@ RSpec.feature "Evidence", type: :system do
     expect(page).to have_content(
       "Select all the categories that describe this file"
     )
-  end
-
-  def when_i_click_back
-    click_on "Back"
   end
 
   def when_i_choose_categories_for_the_second_item
