+++ spec/system/referrals/user_adds_previous_misconduct_spec.rb
@@ -1,13 +1,15 @@
 require "rails_helper"
 
 RSpec.feature "Employer Referral: Previous Misconduct", type: :system do
+  include CommonSteps
+
   scenario "User provides details of previous misconduct" do
     given_the_service_is_open
     and_i_am_signed_in
     and_the_employer_form_feature_is_active
     and_the_user_accounts_feature_is_active
     and_i_have_an_existing_referral
-    and_i_am_on_the_referral_summary_page
+    and_i_visit_the_referral
     when_i_click_on_previous_misconduct
     then_i_see_the_previous_misconduct_reported_page
 
@@ -48,33 +50,20 @@ RSpec.feature "Employer Referral: Previous Misconduct", type: :system do
 
     when_i_choose_no_come_back_later
     and_i_click_save_and_continue
-    then_i_am_on_the_referral_summary_page
+    then_i_see_the_referral_summary
     and_i_see_previous_misconduct_flagged_as_incomplete
 
     when_i_go_back
     and_i_choose_complete
     and_i_click_save_and_continue
-    then_i_am_on_the_referral_summary_page
+    then_i_see_the_referral_summary
     and_i_see_previous_misconduct_flagged_as_complete
   end
 
   private
 
-  def and_i_am_on_the_referral_summary_page
-    visit edit_referral_path(@referral)
-  end
-
-  def and_i_am_signed_in
-    @user = create(:user)
-    sign_in(@user)
-  end
-
   def and_i_choose_complete
     choose "Yes, I’ve completed this section", visible: false
-  end
-
-  def and_i_have_an_existing_referral
-    @referral = create(:referral, user: @user)
   end
 
   def and_i_see_no_previous_misconduct_is_prefilled
@@ -109,24 +98,8 @@ RSpec.feature "Employer Referral: Previous Misconduct", type: :system do
     expect(page).to have_content("upload.txt")
   end
 
-  def and_the_employer_form_feature_is_active
-    FeatureFlags::FeatureFlag.activate(:employer_form)
-  end
-
-  def given_the_service_is_open
-    FeatureFlags::FeatureFlag.activate(:service_open)
-  end
-
-  def and_the_user_accounts_feature_is_active
-    FeatureFlags::FeatureFlag.activate(:user_accounts)
-  end
-
   def then_i_am_asked_to_make_a_choice
     expect(page).to have_content("Tell us if you have completed this section")
-  end
-
-  def then_i_am_on_the_referral_summary_page
-    expect(page).to have_current_path(edit_referral_path(@referral))
   end
 
   def then_i_see_the_details_prefilled
@@ -195,11 +168,6 @@ RSpec.feature "Employer Referral: Previous Misconduct", type: :system do
   def when_i_click_on_previous_misconduct
     click_on "Previous allegations"
   end
-
-  def when_i_click_save_and_continue
-    click_on "Save and continue"
-  end
-  alias_method :and_i_click_save_and_continue, :when_i_click_save_and_continue
 
   def when_i_enter_details_as_text
     choose "I’ll give details of the previous allegations", visible: false
